move GroupController from package controllers to controller and GroupDto from package domain to dto
