+++ src/main/java/com/kodilla/ecommercee/controller/GroupController.java
@@ -1,6 +1,6 @@
 package com.kodilla.ecommercee.controller;
 
-import com.kodilla.ecommercee.domain.GroupDto;
+import com.kodilla.ecommercee.dto.GroupDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
